로그아웃 구현/수정과 삭제 해당 ID만 가능하게 수정

+++ polls/views.py
@@ -9,10 +9,11 @@ from django.urls import reverse
 from django.utils import timezone
 from django.contrib import messages
 
-id = '###'
+g_id = 'dudududu'
 
 def home(request):
-    if id == '###':
+    g_id = 'dudududu'
+    if g_id == 'dudududu':
         return render(request, 'polls/login.html')
     else:
         return HttpResponseRedirect(reverse('index'))
@@ -21,6 +22,8 @@ def login(request):
     if User.objects.filter(u_id=request.POST['id']).count() == 1:
         user = User.objects.get(u_id=request.POST['id'])
         if user.u_passwd == request.POST['passwd']:
+            global g_id
+            g_id = user.u_id
             return HttpResponseRedirect(reverse('index'))
     messages.success(request, "로그인실패")
     return HttpResponseRedirect(reverse('home'))
@@ -34,10 +37,13 @@ def register_do(request):
     if User.objects.filter(u_id=request.POST['id']).count() == 1:
         raise Http404("존재하는 아이디입니다.")
     User.objects.create(u_id=request.POST['id'], u_passwd=request.POST['passwd1'])
-#    global id = request.POST['id']
+    global g_id
+    g_id = request.POST['id']
     return HttpResponseRedirect(reverse('index'))
 
 def logout(request):
+    global g_id
+    g_id = 'dudududu'
     return HttpResponseRedirect(reverse('home'))
 
 def index(request):
@@ -63,14 +69,17 @@ def modify(request, post_id):
     post = Post.objects.get(pk=post_id)
     post.post_title = request.POST['post_title']
     post.post_text = request.POST['post_text']
-    post.save()
+    global g_id
+    if g_id == post.post_author or g_id == 'admin':
+        post.save()
     return HttpResponseRedirect(reverse('detail', args=(post.id,)))
 
 def new(request):
     return render(request, 'polls/new.html')
 
 def create(request):
-    Post.objects.create(post_title=request.POST['post_title'], post_text=request.POST['post_text'], created_date=timezone.now())
+    global g_id
+    Post.objects.create(post_title=request.POST['post_title'], post_text=request.POST['post_text'], created_date=timezone.now(), post_author=g_id)
     return HttpResponseRedirect(reverse('index'))
 
 def comment(request, post_id):
@@ -80,5 +89,7 @@ def comment(request, post_id):
 
 def remove(request, post_id):
     post = Post.objects.get(pk=post_id)
-    post.delete()
+    global g_id
+    if g_id == post.post_author or g_id == 'admin':
+        post.delete()
     return HttpResponseRedirect(reverse('index'))
